pyproject.toml
[tool.poetry]
name = "api"
version = "0.1.0"
description = ""
authors = ["Ygor Tavela <ygortavela@gmail.com>"]

[tool.poetry.dependencies]
python = "3.10.5"
fastapi = "0.78.0"
uvicorn = "0.17.6"
nats-py = "2.1.4"
pyyaml = "6.0"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
